refactor: 계층 분리 고도화 & RestaurantItem 컴포넌트 분리
